Split sensu module into library
