+++ Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "shush"
 version = "0.2.3"
+edition = "2018"
 authors = ["John Baublitz <john.m.baublitz@gmail.com>"]
 description = "A management tool for silencing Sensu checks written in Rust"
 repository = "https://github.com/threatstack/shush"
@@ -12,12 +13,16 @@ include = [
   "LICENSE",
 ]
 
+[workspace]
+members = ["sensu"]
+
 [dependencies]
 clap = "2.33.0"
 hyper = "0.12"
 itertools = "0.6.0"
 nom = "5.0"
 serde_json = "1.0.24"
+sensu = {path = "sensu" }
 regex = "1.1.0"
 rust-ini = "0.10.0"
 tokio = "0.1"
