+++ sensu/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "sensu"
+version = "0.1.0"
+authors = ["Chris Holcombe <cholcombe@box.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+hyper = "0.12"
+itertools = "0.6.0"
+regex = "1.1.0"
+serde_json = "1.0.24"
+tokio = "0.1"
