+++ sensu/src/lib.rs
@@ -1,0 +1,10 @@
+//! Sensu API related request and response-parsing logic
+
+pub mod client;
+pub mod endpoint;
+pub mod err;
+pub mod expire;
+pub mod opts;
+pub mod payload;
+pub mod resource;
+pub mod resources;
