--- src/sensu/mod.rs
@@ -1,16 +1,0 @@
-//! Sensu API related request and response-parsing logic
-
-mod client;
-pub use self::client::*;
-
-mod endpoint;
-pub use self::endpoint::*;
-
-mod expire;
-pub use self::expire::*;
-
-mod payload;
-pub use self::payload::*;
-
-mod resource;
-pub use self::resource::*;
